// src/server/config.d.ts
export declare const config: {
    port: number;
    sessionSecret: string;
    publicUrl: string;
    dataFile: string;
    dbFile: string;
    publicDir: string;
    iconsDir: string;
    clientDir: string;
    mediaRoot: string;
    vaultMediaRoot: string;
    googleClientId: string;
    cacheDir: string;
    smtpHost: string;
    smtpPort: number;
    smtpSecure: boolean;
    smtpUser: string;
    smtpPass: string;
    smtpFrom: string;
};
//# sourceMappingURL=config.d.ts.map